LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-audio-element/audio_constructor-expected.txt


PASS Audio constructor 
FAIL No arguments, without new Audio is not a function. (In 'Audio()', 'Audio' is an instance of AudioConstructor)
PASS No arguments, with new 
FAIL Empty string argument, without new Audio is not a function. (In 'Audio("")', 'Audio' is an instance of AudioConstructor)
PASS Empty string argument, with new 
FAIL Non-empty string argument, without new Audio is not a function. (In 'Audio("src")', 'Audio' is an instance of AudioConstructor)
PASS Non-empty string argument, with new 
FAIL Null argument, without new Audio is not a function. (In 'Audio(null)', 'Audio' is an instance of AudioConstructor)
PASS Null argument, with new 
FAIL Undefined argument, without new Audio is not a function. (In 'Audio(undefined)', 'Audio' is an instance of AudioConstructor)
FAIL Undefined argument, with new assert_equals: expected (object) null but got (string) "undefined"
FAIL Extra argument, without new Audio is not a function. (In 'Audio("", throwingObject)', 'Audio' is an instance of AudioConstructor)
PASS Extra argument, with new 
PASS Calling HTMLAudioElement should throw 
PASS Constructing HTMLAudioElement should throw 
